refactor(workbench-client-testing-app): move application config to separate file
